requirements.txt
requests
faster-whisper
opencc-python-reimplemented
